Organized more event names and map list
Got the rest of the event names corraled. (Except the effects spawning map. I'll leave that one alone.) I also organized the map list and the names of said maps.

IMPORTANT NOTE!!
You cannot move the maps in the list up and down willynilly. The only method I know of so far is to drag it to the top (thing with the box that says the game title) and that will send the map all the way down to the bottom of the list. I don't know why rpg maker made this startlingly bad idea but it is what it is I suppose.

Most important of all:
-Renamed EVT-referenceoutfitchanger to EVT-makegeorgieacripple